doc: alter storybook folders
