+++ .github/workflows/pr.yml
@@ -7,7 +7,7 @@ on:
 
 jobs:
   build:
-    name: Lint & Test
+    name: Build & Test
     runs-on: ubuntu-latest
 
     permissions:
